requirements.txt
langchain>=1.0.0
langgraph>=1.0.0
langchain-anthropic>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
langgraph-checkpoint-sqlite>=2.0.0
langsmith>=0.2.0
# Bedrock support (enterprise) — uncomment when switching LLM_PROVIDER=bedrock
# langchain-aws>=0.2.0
# boto3>=1.35.0
